backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
django_csp==3.8
django-environ==0.11.2
django-templated-mail==1.1.1
pandas==2.0.1
dj-database-url==2.1.0
gunicorn==20.1.0
gevent==24.2.1
greenlet==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.0.0
sqlparse==0.4.3
numpy==1.24.3
contentful==2.2.0
